backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.28.1
PyJWT[crypto]
cryptography
bcrypt==3.2.2
python-multipart==0.0.6
docker==7.0.0
requests==2.31.0
aiofiles==23.2.1
email-validator


